Rename `log/` JavaScript directory to `logs/`
This is going to fuck up the git/GitHub history, but I don't care. It's
worth it. This app has evolved to be about _logs_, plural.
