+++ app/javascript/logs/components/data_renderers/duration_timeseries.vue
