+++ app/javascript/logs/components/data_renderers/integer_timeseries.vue
