+++ app/javascript/logs/components/data_renderers/text_log.vue
