+++ app/javascript/logs/components/log.vue
