+++ app/javascript/logs/components/log_selector.vue
