+++ app/javascript/logs/components/logs_index.vue
