+++ app/javascript/logs/components/new_log_entry_form.vue
