+++ app/javascript/logs/components/new_log_form.vue
