+++ app/javascript/logs/logs.vue
