+++ app/javascript/logs/store.js
